Create StockGrabber.py
Added what I have so far, it takes a certain number of submissions and returns a list of all tickers mentioned (if they used a dollar sign in front of the ticker) and it will output it to a csv. Now I'm working on parsing through comments as well to include those in the list of tickers.